Restructure into one type per file, add upload part size, first review pass
Splits the cmdlets, parameters, and provider into one type per file and moves the
provider partials under a Provider folder. Adds a configurable upload part size and
support for pipeline-bound Set-Content writer parameters. Folds in the first round of
review feedback (capability checks, region fallback, pagination, drive safety) and
makes the tests run headless in CI, tagging the expensive live tests so they stay out
of the daily pipeline.
