+++ modules/AWSPowerShell/Cmdlets/S3/Drive/Provider/S3Provider.Credentials.cs
@@ -1,0 +1,146 @@
+/*******************************************************************************
+ *  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using Amazon;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+
+namespace Amazon.PowerShell.Cmdlets.S3
+{
+    public sealed partial class S3Provider
+    {
+        // Resolves region with the same precedence as the S3 cmdlets (IAWSRegionArgumentsMethods.TryGetRegion):
+        // -Region -> $StoredAWSRegion -> named-profile region -> [default]/legacy profile -> AWS_REGION env
+        // var -> EC2 instance metadata -> us-east-1 (the S3 cmdlets' _DefaultRegion).
+        private RegionEndpoint ResolveRegion(S3DriveParameters dp)
+        {
+            if (!string.IsNullOrEmpty(dp?.Region))
+                return ValidRegionOrThrow(dp.Region);
+
+            var sessionRegion = SessionDefaultRegionName();   // $StoredAWSRegion, via Common
+            if (!string.IsNullOrEmpty(sessionRegion))
+                return ValidRegionOrThrow(sessionRegion);
+
+            // The named profile's own region. StandaloneRegionArguments below can't carry a profile name,
+            // so it only covers the [default]/legacy profile; look up the named one first.
+            if (!string.IsNullOrEmpty(dp?.ProfileName)
+                && SettingsStore.TryGetProfile(dp.ProfileName, null, out var profile)
+                && profile.Region != null)
+                return profile.Region;
+
+            // [default]/legacy profile region, then AWS_REGION, then EC2 instance metadata.
+            var region = new StandaloneRegionArguments().GetRegion(useSDKFallback: true, SessionState);
+            return region ?? RegionEndpoint.USEast1;
+        }
+
+        // GetBySystemName does NOT throw on an unknown name - it returns a synthetic "Unknown" endpoint
+        // that only fails later with an obscure DNS/signing error. Reject it up front instead.
+        private static RegionEndpoint ValidRegionOrThrow(string systemName)
+        {
+            var region = RegionEndpoint.GetBySystemName(systemName);
+            if (string.Equals(region.DisplayName, "Unknown", StringComparison.OrdinalIgnoreCase))
+                throw new ArgumentException(
+                    $"'{systemName}' is not a known AWS region. Use a region system name such as 'us-east-1'.");
+            return region;
+        }
+
+        // Credential precedence: explicit keys -> -ProfileName -> -AWSCredential -> session default
+        // $StoredAWSCredentials (via Get-AWSCredential) -> null (SDK default chain).
+        private AWSCredentials ResolveCredentials(S3DriveParameters dp)
+        {
+            if (dp != null)
+            {
+                // AccessKey and SecretKey are a pair, and SessionToken is meaningless on its own. A
+                // partial set is a mistake (e.g. a typo'd parameter name), so fail rather than silently
+                // dropping to profile/session credentials and mounting against the wrong account. Matches
+                // Set-AWSCredential, which rejects an access key without a secret key (and vice versa).
+                var hasAccessKey = !string.IsNullOrEmpty(dp.AccessKey);
+                var hasSecretKey = !string.IsNullOrEmpty(dp.SecretKey);
+                if (hasAccessKey != hasSecretKey)
+                    throw new ArgumentException(
+                        "-AccessKey and -SecretKey must be supplied together.");
+                if (!string.IsNullOrEmpty(dp.SessionToken) && !hasAccessKey)
+                    throw new ArgumentException(
+                        "-SessionToken requires -AccessKey and -SecretKey.");
+
+                if (hasAccessKey)
+                {
+                    return string.IsNullOrEmpty(dp.SessionToken)
+                        ? (AWSCredentials)new BasicAWSCredentials(dp.AccessKey, dp.SecretKey)
+                        : new SessionAWSCredentials(dp.AccessKey, dp.SecretKey, dp.SessionToken);
+                }
+
+                if (!string.IsNullOrEmpty(dp.ProfileName))
+                    return ResolveProfile(dp.ProfileName);
+
+                if (dp.AWSCredential != null)
+                    return dp.AWSCredential;
+            }
+
+            return SessionDefaultCredentials();   // $StoredAWSCredentials, via Common; null if unset
+        }
+
+        private static AWSCredentials ResolveProfile(string profileName)
+        {
+            var chain = new Amazon.Runtime.CredentialManagement.CredentialProfileStoreChain();
+            if (chain.TryGetAWSCredentials(profileName, out var creds))
+                return creds;
+            throw new ArgumentException($"AWS profile '{profileName}' was not found.");
+        }
+
+        // ---- AWS.Tools.Common session defaults ----
+        // Read through the Common cmdlets rather than the raw session variables: $StoredAWSCredentials
+        // is an AWSPSCredentials whose inner AWSCredentials is internal to Common, so Get-AWSCredential
+        // does the unwrap for us. Best-effort: any failure (Common not loaded, no default set) falls
+        // through to the next resolution step.
+
+        // $StoredAWSCredentials, unwrapped to raw AWSCredentials by Get-AWSCredential. Null if unset.
+        private AWSCredentials SessionDefaultCredentials()
+        {
+            try
+            {
+                var results = SessionState.InvokeCommand.InvokeScript("Get-AWSCredential");
+                foreach (var o in results)
+                    if (o?.BaseObject is AWSCredentials creds)
+                        return creds;
+            }
+            catch { /* Common absent or cmdlet failed -> fall through to the SDK default chain */ }
+            return null;
+        }
+
+        // $StoredAWSRegion system name via Get-DefaultAWSRegion. Null if unset. The cmdlet writes an
+        // AWSRegion whose .Region property holds the system name.
+        private string SessionDefaultRegionName()
+        {
+            try
+            {
+                var results = SessionState.InvokeCommand.InvokeScript("Get-DefaultAWSRegion");
+                foreach (var o in results)
+                {
+                    if (o == null) continue;
+                    var prop = o.Properties["Region"];
+                    var name = prop?.Value as string;
+                    if (!string.IsNullOrEmpty(name)) return name;
+                }
+            }
+            catch { /* Common absent or cmdlet failed -> fall through to us-east-1 */ }
+            return null;
+        }
+
+    }
+}
